refactor(shared/api): rewrite with public exports
