+++ apps/frontend/src/shared/api/index.ts
@@ -1,0 +1,4 @@
+export { ApiError } from "./errors";
+
+export type { Primitive, QueryParams } from "./http";
+export { http } from "./http";
