+++ apps/frontend/src/shared/index.ts
@@ -1,0 +1,4 @@
+export type { Primitive, QueryParams } from "./api";
+export { ApiError, http } from "./api";
+
+export { APP_NAME } from "./constants";
